Markdown Support
Implemented Markdown support for annotations. Originally tried to use the built-in
Markdown plugin for annotatorjs which relied on Showdown but I struggled to get
it working and ultimately decided I only need quite limited support. As a result
I created a custom extension (copied from theirs) which relies instead on
Micromarkdown. This happens to close #2.