Changing font family
It appears that preferring Courier New to the basic monospace font can
cause text to display in a very light colour on Ubuntu and others. As a result
I've changed it to prefer Liberation Mono, then monospace and then Courier New.
Apparently Courier New is default monospace on Windows anyway.